oc_folders: menu usability improvements
This will (re-)show the actions menu after attaching or detaching folders or changing locks. It will assume the attach or detach succeeded as far as calculating the state for the menu is concerned, but one can easily re-calculate by going out of the menu and back in. This change helps with troubleshooting and helps mitigate lag and afk avatar issues.